fix: harden bounded diagnostics and support packaging
