+++ src/app.rs
@@ -13,6 +13,7 @@ use std::sync::{Arc, Mutex};
 
 use serde_json::Value;
 
+use crate::diagnostics::append_json_line as append_line;
 use crate::exitprobe::runtime::{observe_exit, ExitProbeRuntime};
 use crate::exitprobe::{NodeRef, ProbeLimits};
 use crate::georesolver::runtime::{resolve_geo, GeoProviderRuntime};
@@ -26,11 +27,6 @@ use crate::service::status::{
 };
 use crate::service::supervisor::{SupervisorError, UnifiedSupervisor};
 use crate::wloc::PatchTarget;
-
-/// Upper bounds for the root-local structured event log. These are small
-/// enough for AX6S-class gateways while retaining recent diagnostic history.
-pub(crate) const MAX_EVENT_LOG_BYTES: usize = 64 * 1024;
-pub(crate) const MAX_EVENT_LINE_BYTES: usize = 2048;
 
 fn current_unix() -> u64 {
     std::time::SystemTime::now()
@@ -725,50 +721,6 @@ fn probe_needed(
     !fresh || current_fingerprint != last_fingerprint
 }
 
-/// Append one JSON line to an append-only log file (bounded to avoid
-/// unbounded growth).
-fn append_line(path: &std::path::Path, value: &serde_json::Value) {
-    let mut line = serde_json::to_string(value).unwrap_or_default();
-    line.push('\n');
-    if line.len() > MAX_EVENT_LINE_BYTES {
-        return;
-    }
-    if let Some(parent) = path.parent() {
-        let _ = std::fs::create_dir_all(parent);
-    }
-
-    let existing = std::fs::read(path).unwrap_or_default();
-    if existing.len().saturating_add(line.len()) <= MAX_EVENT_LOG_BYTES {
-        if let Ok(mut file) = std::fs::OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(path)
-        {
-            use std::io::Write as _;
-            let _ = file.write_all(line.as_bytes());
-        }
-        return;
-    }
-
-    let keep_bytes = MAX_EVENT_LOG_BYTES.saturating_sub(line.len());
-    let mut retained = if existing.len() > keep_bytes {
-        let start = existing.len() - keep_bytes;
-        let boundary = existing[start..]
-            .iter()
-            .position(|byte| *byte == b'\n')
-            .map(|offset| start + offset + 1)
-            .unwrap_or(existing.len());
-        existing[boundary..].to_vec()
-    } else {
-        existing
-    };
-    retained.extend_from_slice(line.as_bytes());
-    let temporary = path.with_extension(format!("log.tmp.{}", std::process::id()));
-    if std::fs::write(&temporary, retained).is_ok() {
-        let _ = std::fs::rename(temporary, path);
-    }
-}
-
 #[cfg(test)]
 mod probe_needed_tests {
     use super::probe_needed;
@@ -800,7 +752,8 @@ mod probe_needed_tests {
 
 #[cfg(test)]
 mod bounded_log_tests {
-    use super::{append_line, MAX_EVENT_LINE_BYTES, MAX_EVENT_LOG_BYTES};
+    use super::append_line;
+    use crate::diagnostics::{MAX_EVENT_LINE_BYTES, MAX_EVENT_LOG_BYTES};
     use serde_json::json;
 
     #[test]
